fix: Move FetureChanged dialogs to Activity
